img2table_test_project/requirements.txt
img2table==1.2.6
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0
PyMuPDF>=1.18.0
numpy>=1.21.0
openpyxl>=3.0.0
requests>=2.25.0